feat: added support for fastify

+++ src/interceptors/FormData.interceptor.ts
@@ -1,4 +1,10 @@
-import { CallHandler, ExecutionContext, Inject, Injectable, NestInterceptor } from '@nestjs/common';
+import {
+  CallHandler,
+  ExecutionContext,
+  Inject,
+  Injectable,
+  NestInterceptor,
+} from '@nestjs/common';
 import { Reflector } from '@nestjs/core';
 import { FORM_DATA_REQUEST_METADATA_KEY } from '../decorators/form-data';
 import { FormDataInterceptorConfig } from '../interfaces/FormDataInterceptorConfig';
@@ -9,26 +15,36 @@ import { GLOBAL_CONFIG_INJECT_TOKEN } from '../config/global-config-inject-token
 import { checkConfig } from '../helpers/check-config';
 import { is } from 'type-is';
 
-
 @Injectable()
 export class FormDataInterceptor implements NestInterceptor {
   reflector: Reflector = new Reflector();
 
-  constructor(@Inject(GLOBAL_CONFIG_INJECT_TOKEN)
-              private globalConfig: FormDataInterceptorConfig) {
+  constructor(
+    @Inject(GLOBAL_CONFIG_INJECT_TOKEN)
+    private globalConfig: FormDataInterceptorConfig,
+  ) {}
 
-  }
+  async intercept(
+    context: ExecutionContext,
+    next: CallHandler<any>,
+  ): Promise<Observable<any>> {
+    const httpRequest = context.switchToHttp().getRequest();
 
-
-  async intercept(context: ExecutionContext, next: CallHandler<any>): Promise<Observable<any>> {
-    const req = context.switchToHttp().getRequest();
+    /**
+     * fastify always have raw property in the request
+     */
+    const isFastify = !!httpRequest.raw;
+    const req = isFastify ? httpRequest.raw : httpRequest;
 
     /** if the request is not multipart, skip **/
-    if (!is(req, ['multipart'])) return next.handle();
+    if (!is(httpRequest, ['multipart'])) return next.handle();
 
     /** merge global config with method level config **/
     const config: FormDataInterceptorConfig = checkConfig(
-      this.reflector.get(FORM_DATA_REQUEST_METADATA_KEY, context.getHandler()) || {},
+      this.reflector.get(
+        FORM_DATA_REQUEST_METADATA_KEY,
+        context.getHandler(),
+      ) || {},
       this.globalConfig,
     );
 
@@ -36,21 +52,19 @@ export class FormDataInterceptor implements NestInterceptor {
 
     return from(formReader.handle()).pipe(
       mergeMap((formReaderResult: any) => {
-        req.body = formReaderResult;
+        httpRequest.body = formReaderResult;
+
         return next.handle();
       }),
 
-      catchError(err => {
-        if (config.autoDeleteFile)
-          formReader.deleteFiles();
+      catchError((err) => {
+        if (config.autoDeleteFile) formReader.deleteFiles();
         return throwError(err);
       }),
 
-      tap(res => {
-        if (config.autoDeleteFile)
-          formReader.deleteFiles();
+      tap((res) => {
+        if (config.autoDeleteFile) formReader.deleteFiles();
       }),
     );
-
   }
 }
